+++ backend/Controllers/BestellingController.cs
@@ -187,56 +187,3 @@ public class BestellingController : ControllerBase
     }
 }
 
-public class BestellingBody
-{
-    public int Totaalbedrag { get; set; }
-    public string BestelDatum { get; set; }
-
-    public double Kortingscode { get; set; }
-
-    public List<string> Stoelen { get; set; }
-}
-
-public class BestelInfo{
-    public int ShowId {set; get;}
-    public List<int> StoelIds {set; get;}
-    public string AccessToken {set; get;}
-}
-
-
-public class ShowStoelen{
-    public int ShowId {set; get;}
-    public string ShowNaam {set; get;}
-    public string ShowImage {set; get;}
-    public DateTime Datum {set; get;}
-    public List<Stoel> Stoelen {set; get;}
-}
-public class Kaartjeshouder{
-    public string Voornaam {set; get;}
-    public string Achternaam {set; get;}
-    public string Email {set; get;}
-    public List<StoelData> Stoelen {set; get;}
-}
-public class BestellingCleaner{
-    public static async Task Clean(GebruikerContext _context){
-                //Clean up old inactive and unpaid bestellingen
-        bool anyOldActive = await _context.Bestellingen.Where(b => b.BestelDatum < DateTime.Now.AddMinutes(-10)).AnyAsync(b => b.IsActive);
-        if(anyOldActive){
-            var OldActiveBestellingen = await _context.Bestellingen.Where(b => b.BestelDatum < DateTime.Now.AddMinutes(-10)).Where(b => b.IsActive).ToListAsync();
-            foreach (var b in OldActiveBestellingen){
-                b.IsActive = false;
-            }
-            await _context.SaveChangesAsync();
-        }
-        bool anyUnpaidInactive = await _context.Bestellingen.Where(b => b.IsActive == false).AnyAsync(b => b.isBetaald == false);
-        if(anyUnpaidInactive){
-            var UnpaidInativeBestellingen = await _context.Bestellingen.Where(b => b.IsActive == false).Where(b => b.isBetaald == false).ToListAsync();
-            foreach (var b in UnpaidInativeBestellingen){
-                var SeatsToBeRemoved = await _context.BesteldeStoelen.Where(s => s.BestellingId == b.BestellingId).ToListAsync();
-                _context.RemoveRange(SeatsToBeRemoved);
-                _context.Remove(b);
-            }
-            await _context.SaveChangesAsync();
-        }
-    }
-}
