+++ backend/Controllers/KlantController.cs
@@ -138,64 +138,4 @@ public class KlantController : ControllerBase
     }
 
 }
-public class KlantInfo{
-    public bool TwoFactorAuthSetupComplete {set; get;}
-    public bool IsVerified {set; get;}
-    public bool IsBlocked {set; get;}
-    public AccessToken AccessToken {set; get;}
-    public string Voornaam {set; get;}
-    public string Achternaam {set; get;}
-    public string Email {set; get;}
-    public string Beschrijving {set; get;}
-    public string Afbeelding {set; get;}
-    public DateTime? GeboorteDatum {set; get;}
-    public bool IsDonateur {set; get;}
-    public bool IsArtiest {set; get;}
-    public string RolNaam {set; get;}
-    
-}
-public class EmailWachtwoord{
-    public string Email {set; get;}
-    public string Wachtwoord {set; get;}
-}
-public class AccessTokenObject{
-    public string AccessToken {set; get;}
-}
-public class AuthenticatieTokenNieuwWachtwoord{
-    public string AuthenticatieToken {set; get;}
-    public string NieuwWachtwoord {set; get;}
-}
 
-public class NieuweKlant{
-    public string Email {set; get;}
-    public string Wachtwoord {set; get;}
-    public string Voornaam {set; get;}
-    public string Achternaam {set; get;}
-}
-
-public class EmailToken{
-    public string Email {set; get;}
-    public string Token {set; get;}
-}
-public class AccessTokenKey{
-    public string AccessToken {set; get;}
-    public string Key {set; get;}
-}
-
-public class ResponseList{
-        //Custom namen toevoegen aan tuples
-    public static Dictionary<string, (int StatusCode, string Message)> Responses = new Dictionary<string, (int StatusCode, string Message)>(){
-        {"Success", (200, "Success!")},
-        {"AlreadyVerifiedError", (403, "User already verified!")},
-        {"UserNotFoundError", (400, "User not found!")},
-        {"ExpiredTokenError", (403, "Token expired!")},
-        {"NotVerifiedError", (403, "User not verified!")},
-        {"InvalidCredentialsError", (401, "Email or password incorrect!")},
-        {"DisposableMailError", (406, "Disposable email used!")},
-        {"EmailInUseError", (409, "Email in use!")},
-        {"AlreadySetup2FA", (403, "User has already setup their 2FA!")},
-        {"Invalid2FactorKeyError", (401, "Invalid key used!")},
-        {"UserBlockedError", (401, "User has been blocked because of too many login attempts!")},
-        {"TwoFactorNotSetup", (200, "User hasn't set up 2FA")}
-    };
-}
